pyproject.toml
[tool.poetry]
name = "hackin-winters"
version = "0.1.0"
description = ""
authors = ["deadaf <gahlyanrohit18@gmail.com>"]
readme = "README.md"
packages = [{include = "hackin_winters"}]

[tool.poetry.dependencies]
python = "^3.8"
python-decouple = "^3.6"
tortoise-orm = "^0.19.2"
fastapi = "^0.86.0"
asyncpg = "^0.27.0"
uvicorn = "^0.19.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
